Bugfixes for watch_for_replays.ps1
watch_for_replays.ps1 was not uploading files since the $name variable
was not the full filepath. We still want the $name variable to be what
the name of the object is once it is pushed up, but we don't want the
full filepath in the name in object store. So, we move $name to be used
with the --object-name flag in the swift api, and create the $filepath
variable, which will store the full filepath so that api calls to push
replay files work correctly.

Signed-off-by: Cullen Taylor <cullentaylor@outlook.com>