Refactor to use single starcraft_replays dir
Rather than upload each replay to a replays_unprocessed dir, scan it
and append its data to a csv (which can corrupt the dataset), the
processing will now take place on the entire dataset each time.

This will require some more powerful computing resources, but we can
safely assume the processed replay dataset that moves to watson is
sane.